+++ src/Page/Admin/Store/AdminStoreUpdate/index.ts
@@ -1,0 +1,1 @@
+export { default as AdminStoreUpdate } from "./AdminStoreUpdate";
