+++ src/Page/Admin/Store/hooks/useGetStore.ts
@@ -1,5 +1,3 @@
-import { useState } from "react";
-
 import { useParams } from "react-router-dom";
 import { useRecoilValue } from "recoil";
 import { useStoreQuery } from "@/lib/generated/graphql";
@@ -8,35 +6,11 @@ import { userState } from "@/lib/state/userState";
 
 const useGetStore = () => {
   const { accessToken } = useRecoilValue(userState);
-  const [store, setStore] = useState<{
-    name: string;
-    phone: string;
-    code: string;
-    address: string;
-  }>({
-    name: "",
-    phone: "",
-    code: "",
-    address: ""
-  });
 
   const { storeId } = useParams();
-  const storeQuery = useStoreQuery(
-    graphqlReqeustClient(accessToken),
-    {
-      id: Number(storeId)
-    },
-    {
-      onSuccess: (data) => {
-        if (data.store) {
-          const { name, code, address, phone } = data.store;
-          setStore({ name, code, address, phone });
-        }
-      }
-    }
-  );
-
-  return { store, ...storeQuery };
+  return useStoreQuery(graphqlReqeustClient(accessToken), {
+    id: Number(storeId)
+  });
 };
 
 export default useGetStore;
