+++ src/Page/Admin/Store/index.ts
@@ -1,3 +1,3 @@
 export { default as AdminCreateStore } from "./AdminCreateStore";
 export { default as AdminStoreList } from "./AdminStoreList/AdminStoreList";
-export { default as AdminUpdateStore } from "./AdminUpdateStore";
+export { default as AdminStoreUpdate } from "./AdminStoreUpdate/AdminStoreUpdate";
